requirements.txt
Django==5.2
Django-ckeditor==6.7.2
Django-ckeditor-5==0.2.17
Django-crispy-forms==2.4
Django-debug-toolbar==5.2.0
Django-dotenv==1.4.2
Django-environ==0.12.0
Django-social-share==2.3.0
Django-import-export==4.3.7
Django-jazzmin==3.0.1
Django-static-fontawesome==6.5.2.0
Django-taggit==6.1.0
Gunicorn==23.0.0
Pillow==11.2.1
Psycopg2==2.9.10
Python-dotenv==1.1.0
Flake8==7.2.0
Isort==6.0.1
